// src/lib/engine/lib/yaneuraou.module.d.ts
/// <reference types="emscripten" />

export interface YaneuraOuModule extends EmscriptenModule {
  addMessageListener: (listener: (line: string) => void) => void
  removeMessageListener: (listener: (line: string) => void) => void
  postMessage: (command: string) => void
  terminate: () => void
  ccall: typeof ccall
  FS: typeof FS
}